# horoscope-api/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Horoscope API",
    "version": "1.0.0",
    "description": "Simple Horoscope API with JWT auth"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "AuthSignupRequest": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password",
          "birthdate"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password"
          },
          "birthdate": {
            "type": "string",
            "example": "1995-08-17"
          }
        }
      },
      "AuthLoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        }
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "User signup (auto-detect zodiac)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthSignupRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "409": {
            "description": "Email already registered"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "User login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/horoscope/today": {
      "get": {
        "summary": "Get today horoscope",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/horoscope/history": {
      "get": {
        "summary": "Get last 7 days of horoscopes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  }
}